feat: add useCreateContracts hook
